pneumonia-api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
tensorflow==2.15.0
opencv-python-headless==4.9.0.80
Pillow==10.3.0
numpy==1.26.4
matplotlib==3.9.0
